+++ SteveClient.Engine/Networking/Packets/ClientBound/Play/EntityPositionPacket.cs
@@ -6,7 +6,7 @@ namespace SteveClient.Engine.Networking.Packets.ClientBound.Play;
 public class EntityPositionPacket : ClientBoundPacket
 {
     public int EntityId;
-    public Vector3d Delta;
+    public Vector3i Delta;
     public bool OnGround;
     
     public override void Read(in InPacketBuffer packetBuffer)
